Move chemical unit model to Bims
