+++ sass/admin.py
@@ -13,7 +13,6 @@ from sass.models import (
     TaxonAbundance,
     SassValidationStatus,
     SiteVisitTaxon,
-    Chem,
     SiteVisitChem,
     SassEcologicalCategory,
     SassEcologicalCondition,
@@ -129,14 +128,6 @@ class SiteVisitTaxonAdmin(admin.ModelAdmin):
     )
 
 
-class ChemAdmin(admin.ModelAdmin):
-    list_display = (
-        'chem_code',
-        'chem_description',
-        'chem_unit',
-    )
-
-
 class SiteVisitChemAdmin(admin.ModelAdmin):
     list_display = (
         'site_visit',
@@ -199,7 +190,6 @@ admin.site.register(TaxonAbundance, TaxonAbundanceAdmin)
 admin.site.register(SiteVisitBiotopeTaxon, SiteVisitBiotopeTaxonAdmin)
 admin.site.register(SassValidationStatus, SassValidationStatusAdmin)
 admin.site.register(SiteVisitTaxon, SiteVisitTaxonAdmin)
-admin.site.register(Chem, ChemAdmin)
 admin.site.register(SiteVisitChem, SiteVisitChemAdmin)
 admin.site.register(SassEcologicalCategory, SassEcologicalCategoryAdmin)
 admin.site.register(SassEcologicalCondition, SassEcologicalConditionAdmin)
