+++ sass/models/__init__.py
@@ -7,7 +7,6 @@ from sass.models.taxon_abundance import *  # noqa
 from sass.models.site_visit_biotope_taxon import *  # noqa
 from sass.models.sass_validation_status import *  # noqa
 from sass.models.site_visit_taxon import *  # noqa
-from sass.models.chem import *  # noqa
 from sass.models.site_visit_chem import *  # noqa
 from sass.models.sass_ecological_category import *  # noqa
 from sass.models.sass_ecological_condition import *  # noqa
